requirements.txt
torch
transformers
Pillow
opencv-python
numpy
tqdm
toml
cairosvg
spandrel
spandrel_extra_arches